PlugLayout : Handle errors computing section summaries
If we don't, the entire layout will be broken. We only suppress ProcessException because that's what we'll get for computation errors getting plug values in the summariser. We don't suppress other exception types because those would be due to logic errors in the summariser itself, which we want to know about.