SetUI : Improve PlugValueWidget popup menu items
- Always create menu items, with a placeholder "No Sets Available" message when there aren't any sets.
- Add surrounding spaces when inserting into a set expression.
- Maintain cursor position when inserting into a set expression.
- Prefer `str.format()` over `%`.

One potential downside of this is that the new text doesn't get immediately flushed to the plug, because I change the code to operate on the text rather than the plug value (which helps with maintaining cursor position and replacing selection). Should I rejig it to edit the plug values directly again?