CodeWidget : Avoid spurious emission of `editingFinishedSignal()`
This was necessary to avoid two bad interactions with the SetExpressionPlugValueWidget :

1. When holding down the delete key, the completion menu would show, and we'd flush the current value to the plug, which would cause us to update the widget text, which would move the cursor to the start, preventing the expected deletion.
2. When entering text, the completion menu would show and the current text would be flushed to the plug value, causing scene updates with a value that was extremely unlikely to be an intended one.

I'm not particularly happy about the new `_emitEditingFinished()` method, but have at least managed to simplify MultiLineTextWidget in other places - in particular, removing an unnecessary event filter. Alternative approaches I considered were :

1. Having SetExpressionPlugValueWidget install an event filter to swallow unwanted FocusOut events before they got to the MultiLineTextWidget. This still represents some kind of coupling with MultiLineTextWidget, just a less explicit one. Rejected on the grounds that something more explicit was better.
2. Having SetExpressionPlugValueWidget connect to its own `editingFinishedSignal()` and `return True` to suppress emission to subsequent slots. This is simpler and more self-contained, but wouldn't prevent emission for any slots connected via `connectFront()`.
3. Moving the Completer stuff to MultiLineTextWidget. This might be fairly logical, but MultiLineTextWidget already has a lot of methods, and for clarity I kindof like the separation of stuff that is most likely to be code-specific into CodeWidget. This is probably the best of the alternative options though.

I also attempted a more wide-ranging fix : not emitting `editingFinishedSignal()` when losing focus to _any_ popup menu. This was actually quite simple to implement by checking for `lineEdit.hasFocus()` because in this circumstance it continued to return `True` even after `FocusOutEvent`. But it caused additional problems; certain plug popup menus assume that the plug value and the widget text are in sync when the menu is shown. One example is the "Copy Value" item, which gets the value from the plug. It would be extremely confusing if that copied a different value than was visible in the widget. I abandoned this approach on the grounds that it was too risky. I think possibly this might make in terms of semantics too - the completion menu surely doesn't represent the end of text editing, but the generic popup menu _might_, because we don't really know what's in there.