ShaderPlug : Add `parameterSource()`
This opens the way for the `ParameterInspector` to be able to inspect
any parameter within a shader network. Because node names and shader
names are deduplicated independently, the best way to figure out what
node generated a `ShaderNetwork` handle is build the `ShaderNetwork` and
query the `NetworkBuilder` internal shader map.

Future optimizations could be made to only generate the network elements
that are needed for this query (for example, not populating parameter
values).