CodeWidget : Add read-only background colour
This matches the read-only background colour for single-line fields, which is a little odd in that it is brighter than the background we use when editable. Still, this seems the best compromise until we can do a wider overhaul of the stylesheet to get everything lined up better.

Also added a special case to keep the PythonEditor as it was before, although I'm not sure if this is just because we've grown accustomed to how it was, or if it actually is better like this.