MultiLineTextWidget : Render text faded when disabled
For some reason, Qt only allows us to to query the background colour from the palette in `paintEvent()` if it isn't partially transparent. Out stylesheet was inconsisteny here, using an opaque colour for the regular background and a transparent one for the read-only background. Tidied that up to use opaque for both and added some comments justifying that change. In future the stylesheet is in need of some simplification...