MultiLineTextWidget : Add "activation hint" overlay
This hopefully makes the <kbd>Ctrl</kbd>+<kbd>Return</kbd> much more discoverable. I've deliberately not added methods to enable/disable this functionality as I think it is best applied consistently throughout the entire UI. As soon as you show it in one place and not another, it implies that the shortcut doesn't exist in both.