SceneNode : Don't append hash to itself
Not really a problem, but it was weird that we used h as the initial value of the reduce, and then also appended the result to h.

The reasonable option would be either to use h as an initial value, and then finish with h = reduction, or use a default hash as the initial value, and then append to the reduction. I've done the latter.

I haven't actually tested whether appending h to itself degrades the quality of the hash ... some hash functions do suffer from an issue where the possible outputs of `h.append( h )` don't cover the entire possible hash space, so you lose about 1 bit of randomness when you do this.