SceneWriter : Fix writing of locations with names not valid in USD
In this case, the `USDScene::child( CreateIfMissing )` call renames the child to meet USD naming constraints, so we can't later call `USDScene::child()` with the original name to get the same interface. USDScene only sanitises the name to USD requirements when creating it, _not_ when just reading it. I believe the intention behind that is that it should always be an error to request a child not returned by `USDScene::childNames()`. That does seem reasonable, particularly when the sanitisation isn't reversible.

A better fix might be for USDScene to use a bidirectional transcoding between external names and internal USD identifiers. Then we could always encode the names in `USDScene::child()` and decode them in `USDScene::childNames()`. The only names visible to client code would be the expected ones, and all the mess would be hidden within USDScene itself. That will be trivial to achieve if the following proposal is adopted in USD :

https://github.com/NVIDIA-Omniverse/USD-proposals/blob/bidirectional_transcoding_identifiers/proposals/transcoding_invalid_identifiers/README.md.