SetUI : Fix Spreadsheet popup menus
These were broken in f52a358204d632d2c3777a62ed8e6b54a4d9d6c0, when I changed the menus so that they modified the text in the widget rather than editing the plug value directly. I still think that was the right thing to do - when building a complex expression the usage of the menu doesn't correspond to the completion of editing, so we should wait for the user to explicitly commit the changes when they are done. And by using `insertText()` rather than `setValue()` we get better interactions with things like the undo in the text widget.

The problem is that the Spreadsheet UI makes an _invisible_ dummy PlugValueWidget in order to trick the system into making a popup menu. In this case inserting text is useless : the user can't control where it is inserted and it will never get committed to the plug anyway. So we hack around that with special cases for when the widget is invisible, appending onto the current value and calling `Plug.setValue()` directly.

It seems slightly questionable how useful this is for the `acceptsSetExpression` case, because the insertion position is so key in determining the semantics. But it is at least possible to build up simple unions of sets this way. And the new behaviour is better than the original since it at least places whitespace between set names.