FramingConstraint : Fix crash when applied to non-camera locations
The crash was caused by calling `copy()` on a null pointer obtained from `runTimeCast()`. A call to `runTimeCast()` should always be accompanied by a null check, because the whole point is that it might fail.

I've also tidied up an unnecessary second copy that we were taking when we wanted to modify the camera : since we copied it already, we don't need to copy it again. Technically we could omit even the first copy when dealing with perspective cameras and not extending the clipping planes, but I don't think the extra fiddliness that would entail is worth it.