ShaderUI : Fix poor scaling in _ShaderPath
The `ShaderNetwork.shaders()` function builds a Python dictionary containing copies of all the shaders in the network. And in several places, we were calling that at least once per shader, giving us extremely poor scaling for certain large production networks with around 700 individual shaders. This was bad enough that some users reported avoiding the dialogue entirely.

This commit replaces calls to `shaders()` with individual calls to `getShader()`. That gives around a 300x speedup when walking the entire _ShaderPath hierarchy for a problematic network and querying the `shader:inputs` property for each parameter found. This makes the shader dialogues for ShaderTweaks and ShaderQuery usable again.