OpenColorIO : Defer expansion of context variables in OCIO vars
Previously, the OpenColorIOConfigPlug was baking in values from the current context at the point the OCIO plugs were edited on the ScriptNode. This meant that later changes to the context variable were not reflected in the OCIO variable. Among other things, this meant that the Viewer LUT would fail to
update when changing shot in certain multi-shot workflows.

Moving the variable expansion to `OpenColorIOAlgo::currentContextAndConfig()`
fixes this problem, but also enables more advanced workflows. For instance, a Wedge or ContactSheet node can now set the variable being referenced, and the OCIO config used by OpenColorIOTransform nodes will reflect that change dynamically as the graph executes.

There is one knock-on effect from this, as demonstrated in the test changes. On Windows, we must now use `/` separators in the context variables where before we could get away with `\`. We think it's unlikely that anyone is creating contexts manually like this in Windows - much more likely is usage of OpenColorIOContext and OpenColorIOConfigPlug where `/` has always been a requirement. So we think the very small risk of disruption is worth it for the benefits provided.