SceneReader : Avoid computation error for empty UsdVolumes
In this case, `USDScene::hasObject()` returns true because it just does a shallow check that the prim type is one we know how to load. But when we actually come to load it, USDScene discovers that there is no VDB file actually being referenced, and so is forced to return `nullptr`. To deal with this in IECoreUSD we'd have to extend the ObjectAlgo API to allow deeper `hasObject()` checks. While that might be worthwhile in the long run, it also makes for more expensive checks. So it seems worthwhile having a cheap and cheerful backdrop in the SceneReader itself.