AnnotationsGadget : Fix crashes if gadget outlives GraphGadget
This is highly unlikely in practice at this very moment. But in an upcoming commit we're going to be doing some threading shenanigans that means that it is possible if a GraphEditor is destroyed while the AnnotationsGadget does work on a background thread. The code is also simplified a little by having an `annotations()` method compared to an error-prone `graphGadget()` method.