Cycles Renderer : Support batch rendering to `ieDisplay`
This still doesn't support a mix of `ieDisplay` and file outputs, but that isn't a particularly common configuration at least. The output handling could benefit from a much bigger overhaul but here I've limited myself to the most minimal changes necessary to fix the problem. Future improvements could include :

- Moving more logic in CyclesOutput, instead of having much of it in `updateOutputs()`.
- A single `ccl::OutputDriver` subclass that supports writing to file via OIIO as well as rendering to `ieDisplay`.

That can wait though.

With this fix we can now use Batch renders where we logically want to in the unit tests instead of using Interactive renders just to use the display driver we want. I'm also pretty sure this fixes #5905, but can't say with 100% certainty because the issue description is so vague.