CodeWidget : Dispose of completion menu on activatedSignal
This fixes the "`Ctrl`+`Return` doesn't always commit my changes" issue reported recently. If the completion menu was visible while `Ctrl`+`Return` was pressed changes would not be committed due to the loss of focus workaround in `CodeWidget._emitEditingFinished()`. Or in the case of a widget in a PlugPopup, the changes would not be committed and the popup would dismiss...

This was particularly apparent with overlapping set names such as "LGT" and "LGT:ALL" and entering the shorter "LGT" set into a SetExpressionPlugValueWidget in a PlugPopup. Typing "LGT" would leave the completion menu up for "LGT:ALL", and then pressing `Ctrl`+`Return` would dismiss the popup without committing the change while often leaving the completion menu still visible.