Tweaks : Remove `allowCreate` metadata
We already allow "Create" mode on all tweaks, which means this metadata was already registered everywhere so there is no need to keep it around. We remove it to simplify the generation of presets on the mode plug, as we were running into issues when forwarding metadata from spreadsheets where the plug `presetNames` was computed wouldn't always have a parent with `tweakPlugValueWidget:allowCreate` registered.