PopupWindow : Add `parent` argument to `popup()` method
This allows the PopupWindow to inherit the display transform from the main UI, and allows `ancestor()` to be used to find things like the main ScriptWindow.

I used `Menu.popup()` as the model for this since it is the closest related API we have. One potential downside is that the parenting isn't performed in Qt, meaning among other things that it won't be visited by `Widget.__propagateDisplayTransformChange`. I think this is probably good enough for now though, as popup windows are ephemeral and the display transform shouldn't change while they are open.