LightEditor : Fix missing icons for mute and solo
In the process of hiding mute and solo icons for light blockers, we also
mistakenly hid them for groups. On further consideration, we also want
to show the icons for light blockers, which will eventually get mute and
solo capabilities. This restores the icons for all locations in the
Light Editor and also fixes an error causing the `filteredLights` plug
to be returned when querying any attribute from a `LightFilter`.