ImageGadget : Convert ImageBuf to 8 bit before upload to OpenGL
We need to conform it to the format that we're telling OpenGL we're providing. Although in theory the ImageBuf might convert automatically using the hint we passed to the constructor, in practice it doesn't seem to. So do a post conversion when necessary. This fixes corruption of 16 bit node icons in the GraphEditor.