Arnold metadata : Show `color_jitter.data_input` in Graph Editor
Arnold doesn't seem to be entirely consistent about whether it declares the `linkable` metadata for all parameters or not, and it wasn't declaring it in this case despite the parameter being linkable in practice. That meant that ArnoldShaderUI fell through to a heuristic that assumes that ints aren't linkable and so don't need to be shown in the GraphEditor. We now declare the `linkable` metadata ourselves, and for good measure, also make `data_input` visible by default rather than hidden behind the `+` button.