Arnold metadata : Define int parameters as linkable
Arnold doesn't seem to be entirely consistent about whether it declares the `linkable` metadata for all parameters or not, and it wasn't declaring it in these cases despite the parameters being linkable in practice. That meant that ArnoldShaderUI fell through to a heuristic that assumes that ints aren't linkable and so don't need to be shown in the GraphEditor. We now declare the `linkable` metadata ourselves.