PathListingWidget : Avoid blocking UI in `setPath( rootPath )`
The root path is the only path that we ever display in fundamental UIs like the HierarchyView and RenderPassEditor. In this case we were needlessly invoking `Path.isLeaf()` and `Path.isValid()` on the UI thread, causing ScenePath and RenderPassPath to trigger arbitrary computation that could block the UI until complete. By taking an early-out for the root path can completely avoid that, with all the remaining work all happening in a BackgroundTask.