UI : Prefer `context()` over `getContext()`
When accessing the context for Editor, NodeToolbar, PlugLayout and PlugValueWidget. This means we are not relying on the compatibility shim, which is now solely for external code.

Note : Here I'm treating GafferCortexUI as "external" and hence haven't touched it - I'm hoping that it is now time for Image Engine to pick up the maintenance for this module.