Viewer : Remove unused context code
This is an example of a failure to synchronise changes from `setContext()` in a higher-level component (the Viewer) to a lower-level component (the toolbar). The context we were passing was literally used for nothing, and would never have been updated if we called `Viewer.setContext()`. Better that this is all gone now.