PlugValueWidget : Emit `DeprecationWarning` for deprecated methods
The alternative API has been available since Gaffer 1.2, so it is time for folks to get up to date. They can always use a warning filter to buy themselves more time if necessary. The PlugValueWidget API is already pretty complex, and removing the legacy layer in future will be a worthwhile simplification.

> Note : Although everything else in the Gaffer codebase has already been updated, a handful of widgets in GafferCortexUI have not. I am hoping that Image Engine will take on this task, and eventually pick up all maintenance of GafferCortex, since it is not part of the public releases.