AttributeInspector : Implement fallbackValue to return inherited values
This enables us to fall back to inspecting the values of inherited attributes, using the fallbackValue approach already used by OptionInspector for default option values registered as metadata. The initial use-case for this is in the LightEditor, but this would also be of use in a future where we include columns displaying attribute values in the Hierarchy View and other Editors.