SetMembershipInspector : Use fallbackValue to differentiate matches
Rather than returning the PathMatcher::Result as the inspected value, we instead implement fallbackValue to allow us to return true from `value()` for an ExactMatch or true from `fallbackValue()` for an AncestorMatch.

This aligns the SetMembershipInspector behaviour more closely with the AttributeInspector, where `fallbackValue()` returns inherited attribute values when no attribute exists at the location, and will help us to reduce the amount of special handling when dealing with each inspector in the future.

We can also simplify the SetMembershipColumn tooltip handling a little here, as InspectorColumn will already be adding "Double-click to toggle" to the tooltip based on the BoolData now returned by the SetMembershipInspector.