AnnotationsGadget binding : Fix GIL management bug
The `annotationText()` method may call `BackgroundTask::cancelAndWait()`, in which case the background task might require the GIL before the cancellation can complete. The lack of GIL management was causing hangs in the unit tests, but is unlikely to have caused any real-world problems, because `annotationText()` is only intended for use in the tests.