LightEditor : Improve display of fallback values
By using the same dimmed foreground colour for fallback values as used by the RenderPassEditor. This does introduce a small amount of additional duplication with the RenderPassEditor's OptionInspectorColumn, but this will be fairly short lived as we'll be replacing both of these columns with a single common InspectorColumn for Gaffer 1.5.