HistoryWindow : Improve presentation of fallback values
Improve the display of fallback values by replacing the SetMembershipInspector specific `_SetMembershipColumn` with a more generic `_ValueColumn` able to present fallback values provided by the inspectors via the new `history:fallbackValue` HistoryPath property. In the process we lose the "green dot" icon presentation of set membership, but we're already not presenting `light:mute` attributes with their "red dot" icon so this seems a reasonable tradeoff. There's a future where we may be able to start configuring this with metadata specifying icons.

There are further improvements that could be made to the HistoryPath itself, such as including entries when values change so it'd be possible to view changes to inherited attributes as part of the history, and making it more apparent that a plug in the history belongs to a disabled node, though these are left to be tackled at another time...