EditScopeUI : Add `Refresh` menu item
This allows the user to get explicit feedback about ContextTracker updates in the unlikely event that they are slow enough to not be done before the menu is accessed. I've deliberately only shown the BusyWidget when the user "opts in" via a refresh as I think most updates are going to be so rapid that it would flicker annoyingly for them.