ColorChooser : Allow negative values for RGBVTMI
We don't really want to allow these, but they can occur naturally in one
component while editing values in another. By accommodating that within the
slider `hardMin`, we get a nice "out of range" triangle to alert us when we've
gone negative, whereas before the slider clamped to zero and gave us a misleading
value. For consistency, we now also allow negative values to be entered directly
in the numeric widgets - if you can get into that state using a slider or numeric
widget for another compoment, why preclude it in the field itself?