GafferSceneUI : Add InspectorColumn
Both the LightEditor and RenderPassEditor have evolved their own versions of this, so we now consolidate into a common InspectorColumn usable by any editor that wishes to add a column using an Inspector to their PathListingWidget.

Later commits will migrate both the LightEditor and RenderPassEditor to use this instead of their internal implementations.