WidgetAlgo : Work around Windows screen capture issues
On Windows capturing a window with a visible popup menu will only include the popup menu if we capture the entire screen and crop. This appears to be due to Qt's `grabWindow()` not supporting the capture of windows with transparent pixels on Windows. GafferUI.PopupWindow sets `QtCore.Qt.WA_TranslucentBackground`, which triggers this case and prevents the popup from being captured along with the parent window. The Qt documentation advises to grab the entire desktop instead...