Docs : Position Spreadsheet submenu relative to menu width
As the menu width can also vary based on platform and desktop environment. This keeps the submenu from overlapping the menu or appearing disconnected from it.

The vertical alignment of the submenu isn't perfect across the gcc9, gcc11 & windows docs builds, but the vertical placement of the submenu seems more forgiving than the horizontal position and we could tidy this up more easily once we drop the gcc9 builds.