PathListingWidget : Add column context menu signal
This follows the pattern established by other context menus, in allowing multiple client slots to collaborate on the building of a context menu.

I debated whether or not the signal should pass the `path` that is currently under the mouse, but decided against. That would encourage the creation of menu items that operate only on that path, but with multiple selection I believe the user expectation would be that the entire selection is operated on. By forcing clients to go fetch the selection themselves they are more likely to do the expected thing with it.