HierarchyView : Add `sceneListing()` method
This is intended to give clients access to the `columnContextMenuSignal()` to customise the context menu. Currently it returns a bare PathListingWidget which gives clients more privilege than is ideal (we don't want them to change the path!), but in future should return a more specialised SceneListing widget. That won't appear till Gaffer 1.5 though, and we want to allow the menu to be customised in 1.4. So for now we content ourselves with a little warning that folks shouldn't get addicted to anything other than the context menu signal.