ShaderUI : Mark `type` plug as read-only
As with `name`, it is already considered read-only because `loadShader()` should be used to set it and load all the associated parameters at the same time. Marking it as read-only via metadata prevents the "Revert to Defaults" menu item from clobbering it, which previously was breaking the shader completely.

Also updated documentation to make use of Markdown code formatting.

Fixes #1862