Editor : Use ContextTracker
The associated Viewer update is sufficient to keep the ImageView and SceneView using the correct context and updating appropriately when it changes. But it reveals a bug in all our Tool subclasses : while these all do dutifully connect to the `changedSignal()` for the new context, they don't do the immediate update that is needed to reflect _this_ change. And of course the `changedSignal()` is never emitted, because the usage of the ContextTracker just means that a _new_ context will arrive, not that the old one will be modified in place. The upshot is that things like transform handles don't update to reflect context changes (like animation). This will be dealt with in future commits.

The Changes.md entry for this isn't particularly informative, and in fact we're just stacking up lots of similar entries for each UI component we touch. I intend to rewrite them all into a more holistic "everything is now focus-aware" entry when I'm done with updating everything.