NodeToolbar : Use ContextTracker
I don't think we actually call `NodeToolbar.context()` from anywhere, but it's part of the public API, and is easy enough to update.

This allows us to finally remove `PlugValueWidget.__defaultContext` which was being used to share some nasty find-the-scriptnode juggling which is now done more centrally in `ContextTracker.acquireForFocus()`.