PathColumn : Allow sorting of FileIconPathColumn
Sorting by file extension is an approximation of sorting by file type, but is likely enough for our purposes.

`std::filesystem::path::extension()` returns the last part of the filename from the rightmost `.`, which means "foo.tar.gz" will sort with "foo.gz", but this is preferable to a more greedy approach that would treat "foo.0001.exr" as having an extension of "0001.exr".