ScriptNodeAlgo : Add new namespace to eventually replace ContextAlgo
Storing UI state in a context (like ContextAlgo does) was a bad idea, and we need to get away from it. The transition plan is this :

1. Introduce ScriptNodeAlgo, implemented on top of ContextAlgo. That's this commit. This will allow clients to transition to the ScriptNodeAlgo API while complete interoperability is maintained with the old ContextAlgo API.
2. Wait until all supported Gaffer versions provide ScriptNodeAlgo (when 1.5.x.x is in bugfix-only maintenance).
3. Remove ContextAlgo, and reimplement ScriptNodeAlgo using Metadata stored on the script.

The reason this has become more pressing is that all UI components are about to be converted to use contexts provided by ContextTracker, so will not typically be working with the ScriptNode context. Many ContextAlgo clients will be broken at that point, because they won't be sharing state in the same shared context. If they use ScriptNodeAlgo instead, that problem won't exist, and we can safely use ContextTracker.