View : Add `scriptNode` constructor argument and `scriptNode()` accessor
Access to a ScriptNode is key for any UI component, because that is where the shared UI state is stored (context, focus node, selection etc). Previously we could only gain access to a ScriptNode when an input connection existed at `View.in`, which was not always guaranteed (for one, the UVInspector's UVView has no input when no scene is being viewed). And even doing this required more hoop jumping than we'd want for such a common operation.

This will allow us to do several things in future :

- Simplify client code needing access to a ScriptNode.
- Remove `View::setContext()` and instead use a ContextTracker, which will soon be the way we do things everywhere in the UI.
- Use a new `GafferSceneUI.ScriptNodeAlgo` API that will replace `GafferSceneUI.ContextAlgo`.

This commit doesn't do any of that. It just adds the constructor argument and fixes up all the relevant call sites to keep things running as before. Although there are a lot of call sites, they are mainly in View/Tool unit tests and View subclasses, and we don't expect there to be too much code affected outside of the Gaffer source itself.