PrimitiveInspector : Fix thread-safety and cancellation bugs
We need to scope the context _before_ calling the background method, rather than from within the background method. This automatically creates a copy of the context which we can safely use from the background thread, and also organises a canceller which will respond to graph edits and the hiding of the PrimitiveInspector.