SetEditor : Prefer ScriptNodeAlgo to ContextAlgo (phase 2)
This moves the handling of selection out of SetPath and into the SetEditorEmptySetFilter and SetSelectionColumn, switching to ScriptNodeAlgo at the same time. It felt a bit more natural to have selection be managed in the "presentation layer" rather than in the data (Path) layer itself. But it would also be possible to have kept the original organisation. Can switch back if preferred.