TypedObjectPlugBinding : Add CompoundData->CompoundObject conversion
This allows `CompoundObjectPlug.setValue( compoundData )` to work, meaning we can remove the `attributesCompatibility.py` shim that was doing this manually for just `Attributes.extraAttributes`.

What I don't like about this is the mismatch between the C++ and Python APIs : it's still not possible to call `CompoundObjectPlug::setValue( CompoundData )` in C++. And that's also mismatched with `CompoundObjectPlug::setFrom()` which is implemented to allow connections from CompoundDataPlug. Still, it seems like progress compared to the compatibility shim.

The immediate motivation for this is that it's not currently possible to have `userDefault` metadata for a CompoundObjectPlug, because metadata values must all be `Data`. This allows the registation of a CompoundData value which is converted to a CompoundObject when applied to the plug.