SignalBinding : Default `scoped = False` for `connect[Front]()`
This has a long history. Originally `connect()` was bound to return `ScopedConnection` rather than the usual `Connection` object returned in C++. This was confusing and inconvenient :

```
signal.connect( mySlot )
connection = signal.connect( mySlot )
```

So in 0.14.0.0 we added a `scoped` argument, which could be used to request the same behaviour as in C++, returning a `Connection` object instead.

 ```
signal.connect( mySlot, scoped = False )
signal.connect( mySlot )
```

Then in 0.54.0.0 we introduced a Trackable Python class that removed the need for scoped connections in many cases (#3179). Unscoped connections were now more prevalent than scoped, but the default was still `scoped = True`. And that default was still causing confusion.

In 0.62.0.0 we paved the way to eventually clearing up this mess. We deprecated the default value for the `scoped` argument, requiring every caller to pass either `scoped = True` or `scoped = False` explicitly (#4535). If they didn't they'd receive a DeprecationWarning, which we treat as a failure in our own unit tests. Now, more than two years and several major versions later, I think we can assume that everyone has updated their code. That puts us in a situation where we can reintroduce a default value, but this time with `scoped = False`. Which means that finally, after all these years, a simple `signal.connect( slot )` no longer automatically disconnects itself immediately.