VectorDataWidget : Disable word wrap
Back in the mists of time a Qt update caused paths displayed in VectorDataWidgets to be more aggressively elided when the path was longer than the widget could display. A path of `/a/longpath` would be elided to `/a/...` rather than `/a/long...`. We suspected some change in behaviour to how Qt was handling elided text, but early investigations proved to be not so fruitful, and the alternative of implementing `_StringDelegate.paint()` in Python didn't feel ideal for performance.

It turns out that we were barking up the wrong tree there and the real issue was word wrapping. The path text was being wrapped before the text was elided. The wrapping considered '/' characters to be appropriate places to introduce line breaks, which are indicated with the same ellipses as elided text when the widget cannot display the additional line(s). Disabling word wrap on the VectorDataWidget's TableView allows the full path to be elided and we return to displaying `/a/long...` rather than `/a/...`.