ContextTracker : Avoid unnecessary reference counting operations
We can't use raw pointers as the keys in these maps, because then the node could be destroyed and replaced by another at the same address, and our queries would return contexts that are completely irrelevant to the new node. This would be alleviated somewhat by the destruction of the old node triggering an update, but since those are asynchronous it would still leave us in a bad state until the update was complete. What we really want is a weak pointer key, but we don't have weak pointers to RefCounted types (although we could benefit from them all over the place).

Because the key is `intrusive_ptr<T>`, by default `find( T * )` will cast the argument to an `intrusive_ptr`, doing a completely pointless reference count increment and decrement. That's a small performance problem that's probably worth avoiding, but there's a bigger problem. It prevents a View from requesting a context for itself during construction, because the reference count would be starting at zero and hitting zero again, destructing the View prematurely. By providing a hash function that can handle both raw and intrusive pointers we can enable "heterogeneous lookup", and hash the raw pointer directly without creating a temporary. We resort to using `boost` for this for now, since `std` doesn't provide heterogeneous lookup until C++20. The `using is_transparent = void` alias is unnecessary for boost, but prepares us for the switch back to `std` in C++20 (at which point we can also remove the extra arguments to `find()`).