ImageGadget : Remove non-const `getContext()` method
And pass const Contexts to `setContext()`. This brings it into line with the SceneGadget, and provides a guarantee that the ImageGadget itself will not modify the context. This makes much more sense for a ContextTracked world, where we're constantly doling out new const Contexts rather than having everything listen directly to a central `ScriptNode::context()` which is modified in place.

The one wrinkle here is that `setContext()` needs a cheeky const cast just to get access to the `changedSignal()` on the context, but since that doesn't break the guarantee that we won't modify the context, I think it's defensible. Connecting to `changedSignal()` is almost redundant now anyway, as we expect most contexts to be coming from a ContextTracker, in which case they won't change, but will just be replaced. We're not there yet, but I can imagine a future where this is the only way of updating the context, and `changedSignal()` disappears.